docs(project): update
